atividade-obrigatoria-18-06-2025/requirements.txt
langchain_community==0.3.25
langchain_openai==0.3.22
pandas==2.3.0
python-dotenv==1.1.0
streamlit==1.45.1
